Refactor soundscapes dataset builder.
Separate into three files:
* soundscapes.py contains all of the beam + tfds logic. (Broadly, anything that requires lots of mocks to test.) We thus ultimately want to minimize the size of this file.
* soundscapes_lib.py contains functionality for processing and combining standardized annotations and metadata.
* dataset_fns.py contains all of the dataset-specific implementations for, eg, reading different annotations+metadata.

Breaking out this functionality makes it much easier to test, and allows creating a bunch of test files which mimic the actual dataset-specific structures, so we can be more sure that we don't break on small changes and can reliably add new datasets in the future.

There is now unit testing for per-dataset functionality and integration testing for combining the various datasources for each builder config. The tests now require some realistic example data for each soundscape builder config.

PiperOrigin-RevId: 472527520