Create a convolutional approximation of EMA for use in PCEN ops.
I tested the new implementation for numerical equivalence on real data; error on 5s PCEN outputs was between ~3e-5 and 1e-3 depending on the sample. A new model trained on the convolutional EMA does fine, though, so seems like an adequate drop-in when used for both train+eval... Though I don't think I would trust it as an inference-only replacement.

PiperOrigin-RevId: 477754795