Xenocanto/IOC 10.1 to ebird2021.
Xeno-Canto is currently using IOC 10.1, which makes conversion a bit difficult. This change provides the data (much of it hand-checked) to convert to ebird2021 species labels. This includes the XC update to IOC 12.2, and conversions from IOC 12.2 to EBird subspecies codes (where applicable).

Since the Xeno-Canto species list is non-static, we provide a generator for a mapping instead of a fixed namespace.

PiperOrigin-RevId: 501039329